[MOV] im_livechat: move Timer outside of public_livechat.js
Task-2212347
